pyproject.toml
[tool.poetry]
name = "configurave"
version = "0.1.0"
description = "A one-size-fits-all configuration scheme manager"
authors = ["Bast <bast@fastmail.com>", "aru <genericusername414@gmail.com>"]
readme = "README.md"
repository = "https://github.com/discord-modmail/configurave"
homepage = "https://github.com/discord-modmail/configurave"

[build-system]
requires = ["poetry-core~=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.7"
tomlkit = "^0.7.2"
python-dotenv = {version = "^0.19.0", optional = true}

[tool.poetry.extras]
dotenv = ["python-dotenv"]

[tool.poetry.dev-dependencies]
black = "^21.7b0"
coverage = { extras = ["toml"], version = "^5.5" }
flake8 = "~=3.8"
flake8-annotations = "~=2.3"
flake8-bandit = "^2.1.2"
flake8-bugbear = "~=20.1"
flake8-docstrings = "~=1.5"
flake8-isort = "^4.0.0"
flake8-string-format = "~=0.3"
flake8-tidy-imports = "~=4.1"
flake8-todo = "~=0.7"
isort = "^5.9.2"
pep8-naming = "~=0.11"
pre-commit = "~=2.1"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
pytest-cov = "^2.12.1"
pytest-dependency = "^0.5.1"
pytest-docs = "^0.1.0"
pytest-sugar = "^0.9.4"
pytest-xdist = { version = "^2.3.0", extras = ["psutil"] }
taskipy = "^1.6.0"


[tool.taskipy.tasks]
black = { cmd = "black --check .", help = "dry run of black" }
flake8 = { cmd = "python -m flake8", help = "Lints code with flake8" }
lint = { cmd = "pre-commit run --all-files", help = "Checks all files for CI errors" }
precommit = { cmd = "pre-commit install --install-hooks", help = "Installs the precommit hook" }
report = { cmd = "coverage report", help = "Show coverage report from previously run tests." }
test = { cmd = "pytest -n auto --dist loadfile --cov-report= --cov= ", help = "Runs tests and save results to a coverage report" }
test-nocov = { cmd = "pytest --no-cov", help = "Runs tests without creating a coverage report" }


[tool.coverage.run]
branch = true
source = ["configurave"]

[tool.pytest.ini_options]
addopts = "--cov -ra"
minversion = "6.0"
testpaths = ["tests"]

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
